% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_repertoire.R
\name{AddCellMetaToVDJ}
\alias{AddCellMetaToVDJ}
\title{Add cell metadata from Seurat object into VDJ data frame}
\usage{
AddCellMetaToVDJ(vdj, SeuratObj, metadata_col, barcode_col = "row.names")
}
\arguments{
\item{vdj}{a data.frame of VDJ data after running \code{\link{collapseBCR}}.}

\item{SeuratObj}{Seurat object.}

\item{metadata_col}{vector of column names from SeuratObj denoting the desired metadata to be extracted and added to vdj}

\item{barcode_col}{column name in SeuratObj metadata which holds cell barcodes. (Default: 'row.names' i.e. use row names of SeruatObj metadata as cell barcodes)}
}
\value{
VDJ data frame with additional columns taken from the Seurat object metadata slot.
}
\description{
\code{AddCellMetaToVDJ} merges annotations from Seurat object (e.g. cluster annotations) into VDJ repertoire data frame.
}
\details{
The function accepts any columns found in the metadata slot of the Seurat object. You need to indicate for each column whether the data
}
