% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{convertSeuratToH5ad}
\alias{convertSeuratToH5ad}
\title{wrapper function to convert Seurat Object to a AnnData .h5ad file}
\usage{
convertSeuratToH5ad(SeuratObj, assays, h5ad_filename)
}
\arguments{
\item{SeuratObj}{Seurat Object}

\item{assays}{A vector of assay names from SeuratObj to be exported}

\item{h5ad_filename}{Filename of the output .h5ad file. Note that the final output filenames will have the assay names appended to this (see examples).}
}
\value{
A vector of .h5ad filenames which are outputted. Each file correspond to one Seurat assay, as indicated in the suffix inside the filename (see examples).
}
\description{
`convertSeuratToH5ad` is a wrapper function to convert a given Seurat Object into an AnnData object (for use in python with e.g. scanpy) and write out into a .h5ad file.
}
\details{
`convertSeuratToH5ad` simply wraps around the R `SeuratDisk` package to perform the stated conversion. Included her for ease of use for the user.
Each assay in the Seurat Object is written into separate .h5ad files.
}
