% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{dim_plot}
\alias{dim_plot}
\title{Plot dimensionality-reduced projection of single-cell data}
\usage{
dim_plot(SeuratObj, group.by = "seurat_clusters", dim_reduce = "UMAP")
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{group.by}{column in Seurat object meta.data to group the cells (Default: 'seurat_clusters')}

\item{dim_reduce}{name of the dimensionality reduction to calculate centroid coordinates (Default: "UMAP")}
}
\value{
ggplot2 object of the dimensionality-reduced projection of the single-cell data, on the first 2 axes
of the projection given in `dim_reduce`
}
\description{
`dim_plot` plots the dimensionality-reduced projection stored in the Seurat object.
}
\details{
`dim_plot` effectively emulates the `DimPlot` function in Seurat, but here returns a native `ggplot2` object
whch enables additional geoms (e.g. the flux network plot from `plotFlux`) to be drawn on top of the dimensionality-reduced
projection.
}
