% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{FindClusterCentroid}
\alias{FindClusterCentroid}
\title{Find the centroid of each cell cluster in the dimensionality-reduced space.}
\usage{
FindClusterCentroid(
  SeuratObj,
  group.by = "seurat_clusters",
  dim_reduce = "UMAP"
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{group.by}{column in Seurat object meta.data to group the cells (Default: 'seurat_clusters')}

\item{dim_reduce}{name of the dimensionality reduction to calculate centroid coordinates (Default: "UMAP")}
}
\value{
a data.frame of the coordinates (in the 2 axes of the dimenesionality reduced space) for each cell cluster.
}
\description{
`FindClusterCentroid` calculates for each cell cluster, the centroid coordinates in a specified dimensionality reduction space.
}
\details{
`FindClusterCentroid` calculates the mean coordinate of the 2 axes of the dimensionality reduction space
specified by `dim_reduce`, for each cluster specified by `group.by`.
}
