% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGHcounts.R
\name{getIGHreadType}
\alias{getIGHreadType}
\title{Deduce type of IGH reads}
\usage{
getIGHreadType(tb)
}
\arguments{
\item{tb}{A data.frame, output from `scanBam`.}
}
\description{
`getIGHreadType` deduces, for each molecule in each cell identified via `scanBam`,
whether it corresponds to a productive (labelled "-P") or a sterile ("-S") IGH
molecule. A third category ("-C") is assigned if insufficient information is present
to distinguish "-P" and "-S".
}
\details{
The function loops through each cell barcode - molecule barcode combination
and looks for evidence of reads mapping to the following three regions: VDJ,
C as well as the 5' intron to C. Each molecule is assigned based on these criteria:
\itemize{
 \item{Productive (-P)} {at least one read mapping to VDJ region and at least one
read mapping to the C exonic region. NOTE: here we are not making the distinction
of whether the VDJ has already been spliced to the C exons at the RNA level. We are
simply asking whether the molecule can encode a Ig protein with both V and C regions.}
 \item{Sterile (-S)} {at least one read mapping to the 5' intronic region to a C gene,
 without reads mapping to the VDJ region. These represent the 'sterile'/'germline'
 IgH transcripts which primes class-switch recombination.}
}
A third category (-C) is assigned if insufficient information is available to classify
the molecule in to the '-P' or the '-S' groups.
}
