% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGHcounts.R
\name{getJunctionReads}
\alias{getJunctionReads}
\title{extract reads covering splice junctions}
\usage{
getJunctionReads(read_info)
}
\arguments{
\item{read_info}{output from `scanBam` containing details of the aligned reads for VDJ, C and intronic regions.}
}
\value{
A data.frame containing details of reads which cover splice junctions:
\describe{
  \item{qname}{read ID in the FASTQ/BAM}
  \item{CB)}{cell barcode}
  \item{UB)}{Unique molecule identifier (UMI)}
  \item{type}{a list of mapped entities (VDJ/IGHC genes/intronic rgeions) found on the read.}
}
}
\description{
`getJunctionReads` extract reads from `scanBam` that cover splice junctions.
}
\details{
`getJunctionReads` extract reads from `scanBam` that cover splice junctions.
This considers only those which cover BOTH VDJ and C exons/introns, and
strange cases which cover multiple different IGHC genes.
}
