% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{getSHM}
\alias{getSHM}
\title{get somatic hypermutation level}
\usage{
getSHM(SeuratObj, v_identity_anno_name, shm_column_to_add = "shm")
}
\arguments{
\item{SeuratObj}{Seurat Object}

\item{v_identity_anno_name}{column name in the Seurat meta.data slot which holds the v_identity information to be considered.}

\item{shm_column_to_add}{name of column to be added to the Seurat meta.data slot
which holds the SHM frequency calculated in this function}
}
\value{
The same Seurat object as given by SeuratObj, except that a new column with
name given by `shm_column_to_add` is appended to the SeuratObj meta.data slot to reflects
the calculated SHM frequency.
}
\description{
`getSHM` calculates somatic hypermutation (i.e. 1 - (percentage identity to germline VH gene)).
}
\details{
`getSHM` considers the `v_identity` (i.e. % identity to germline V gene) for the VH sequence
mapped to each cell, and calculate 1 - `v_identity` as the somatic hypermutation (SHM)
level for the cell. It finds the v_identity information from the meta.data slot of the Seurat object.
For cells without a mapped V sequence, it will impute SHM as 0.
}
