% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_multiple_scales.R
\name{ggplot_add.new_aes}
\alias{ggplot_add.new_aes}
\title{Special behaviour of the "+" for adding a `new_aes` object
It changes the name of the aesthethic for the previous layers, appending
"_new" to them.}
\usage{
\method{ggplot_add}{new_aes}(object, plot)
}
\arguments{
\item{object}{`new_aes` object.}

\item{plot}{`ggplot2` plot object.}
}
\description{
Special behaviour of the "+" for adding a `new_aes` object
It changes the name of the aesthethic for the previous layers, appending
"_new" to them.
}
