% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{guessBarcodes}
\alias{guessBarcodes}
\title{parse the substring inside a given cell identifier which corresponds to the nucleotide barcode}
\usage{
guessBarcodes(cell_name, min_barcode_length = 6L)
}
\arguments{
\item{cell_name}{character, a cell identifier, typicall with prefix and/or suffix (e.g. "ACTGATGCAT-1", "SampleA_ATGAACCTATGG")}

\item{min_barcode_length}{integer, minimum length of the nucleotide barcode (Default: 6)}
}
\value{
a vector with the input \code{cell_name} decomposed into these three entries:
\describe{
  \item{prefix}{prefix which exists in the input \code{cell_name} (\code{NA} if doesn't exist in \code{cell_name})}
  \item{cell_name}{the actual nucleotide barcode}
  \item{suffix}{suffix which exists in the input \code{cell_name} (\code{NA} if doesn't exist in \code{cell_name})}
}
}
\description{
\code{guessBarcodes} parses given cell identifiers to identify the substring which correspond to the nucleotide barcode included in the experiment.
}
\details{
Numeric / string prefices/suffices were typically added to cell identifiers to avoid wrong mapping across samples; however often these create issues when trying
to merge data on the *same* sample but annotated using different workflows.
This function attempts to resolve such issues by extracting the nucleotide barcodes actually introduced in the experiment.
}
