% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_nmf.R
\docType{data}
\name{human_nmf}
\alias{human_nmf}
\title{Isotype sterile/productive signatures trained using human single-cell B cell atlas}
\format{
A matrix with 16 rows and 2 columns, each row corresponds to a IgH sterile/productive gene and each column its weights in the isotype signatures.
}
\usage{
human_nmf
}
\description{
The breakdown of sterile/productive IgH transcripts into isotype 'signatures' using
non-negative matrix factorization (NMF). This object stores the NMF weights of each
sterile/productive IgH gene in each NMf signature.
}
\keyword{datasets}
