% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGHcounts.R
\name{mergeIgHCountsToSeurat}
\alias{mergeIgHCountsToSeurat}
\title{merge IgH productive/sterile transcript count into Seurat Object}
\usage{
mergeIgHCountsToSeurat(igh_counts, SeuratObj, assay_name = "IGHC")
}
\arguments{
\item{igh_counts}{count matrix of productive/sterile transcripts after ensuring cell names/barcodes match with the supplied Seurat Object (i.e. after running the \code{\link{repairBarcode}} function)}

\item{SeuratObj}{Seurat object on to the productive/sterile count matrix will be attached}

\item{assay_name}{name of the Seurat Assay holding the productive/sterile transcript count matrix. (Default: "IGHC")}
}
\value{
A Seurat object identical to \code{SeuratObj} with a new Assay with name given by \code{assay_name}, containing the productive/sterile transcript count matrix \code{igh_counts}.
}
\description{
\code{mergeIgHCountsToSeurat} will generate a new Assay in the given Seurat Object which contains the count matrix for productive/sterile transcripts
}
\details{
\code{mergeIgHCountsToSeurat} expects a count matrix with matching cell names/barcodes with the Seurat Object (i.e. output from the \code{\link{repairBarcode}} of \code{sciCSR}).
It will create a new Assay that stores this count matrix.
}
