% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_definitions.R
\docType{data}
\name{mouse_definitions}
\alias{mouse_definitions}
\title{Genomic coordinates of mouse heavy-chain immunoglobulin V, D, J, C genes}
\format{
A list of three GenomicRanges objects:
\describe{
  \item{VDJ}{GenomicRanges object containing genomic coordinates of individual V, D and J genes}
  \item{C}{GenomicRanges object containing genomic coordinates of the coding segment of individual C genes}
  \item{sterile}{GenomicRanges object containing genomic coordinates of the sterile transcripts of individual C genes}
}
}
\usage{
mouse_definitions
}
\description{
Downloaded from Ensembl BioMart query of the mm10 reference genome. The positions of sterile C transcripts were separately annotated. This is used for enumerating the productive and sterile IgH transcripts from given BAM files.
}
\keyword{datasets}
