% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_nmf.R
\docType{data}
\name{mouse_nmf}
\alias{mouse_nmf}
\title{Isotype sterile/productive signatures trained using mouse single-cell B cell atlas}
\format{
A matrix with 14 rows and 2 columns, each row corresponds to a IgH sterile/productive gene and each column its weights in the isotype signatures.
}
\usage{
mouse_nmf
}
\description{
The breakdown of sterile/productive IgH transcripts into isotype 'signatures' using
non-negative matrix factorization (NMF). This object stores the NMF weights of each
sterile/productive IgH gene in each NMF signature.
}
\keyword{datasets}
