% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{normalise_dimreduce}
\alias{normalise_dimreduce}
\title{Normalisation and dimensionality reduction of scRNAseq gene counts}
\usage{
normalise_dimreduce(
  obj,
  var_explained_lim = 0.015,
  run.harmony = FALSE,
  harmony_vars = NULL,
  SCT = FALSE,
  mt.pattern = "^MT-",
  mt.percent = 10,
  features_exclude = c("^IGH[MDE]", "^IGHG[1-4]", "^IGHA[1-2]", "^IG[HKL][VDJ]", "^IGKC",
    "^IGLC[1-7]", "^TR[ABGD][CV]", "^AC233755.1", "^IGLL", "^JCHAIN"),
  ...
)
}
\arguments{
\item{obj}{Seurat object with the gene counts unnormalised.}

\item{var_explained_lim}{numeric, the minimum proportion of variance explained cutoff for a principal component to be included in the dimensionality reduction and clustering steps (default: 0.015, i.e. 1.5\%)}

\item{run.harmony}{should the package \code{Harmony} be used on the data? (Default: FALSE)}

\item{harmony_vars}{vector of parameters to be included in the regression step in \code{Harmony}. Variations specific to these parameters will be removed during the \code{Harmony} run.}

\item{SCT}{Should the \code{SCTransform} pipeline be used? If not it will follow the standard Seurat normalisation workflow (\code{NormalizeData}, \code{FindVariableFeatures}, \code{ScaleData})}

\item{mt.pattern}{the regular expression used to identify mitochondrial transcripts (Default: ^MT-", i.e. all gene names beginning with "MT-")}

\item{mt.percent}{the cutoff for mitochondrial transcript percentage, above which cells will be removed from the Seurat project as part of quality control (Default: 10, i.e. cells with more than 10\% of counts mapped to mitochondrial transcripts will be removed from the Seurat object)}

\item{features_exclude}{a vector of regular expressions to select genes to be IGNORED during dimensionality reduction and clustering. By default the following features were included in this list: IgH, K, L V/D/JC genes, TRA/TRB V/C genes, AC233755.1 (which encodes a V-gene-like product), IGLL, JCHAIN)}

\item{...}{Arguments to be passed to various Seurat functions (\code{SCTransform}, \code{NormalizeData}, \code{FindVariableFeatures}, \code{ScaleData}, \code{RunPCA}, \code{RunUMAP}, \code{FindNeighbors}, \code{FindClusters})}
}
\value{
A Seurat object with normalised gene count data, dimensionality reduction and clustering done
}
\description{
\code{normalise_dimreduce} is a wrapper function around the Seurat basic data processing workflow to generate normalised gene count data, its dimensionality reduction, and derivation of cell clusters.
}
\details{
\code{normalise_dimreduce} is a wrapper function around the major basic Seurat data processing workflow and performs the following steps:
\itemize{
  \item{Calculation of \% mitochondrial transcripts}{ (\code{PercentageFeatureSet}) and subsetting to remove those beyond the cutoff given by \code{mt.percent}.}
  \item{Gene count normalisation}{, using either \code{SCTransform} or \code{NormalizeData}}
  \item{Pruning variably expressed features}{. All genes with names matching the vector of regular expression given in the argument \code{features_exclude} will be removed from this list to avoid them influencing the downstream dimensionality reduction and clustering steps. This is particularly relevant for avoiding clusters of B cells grouped by their isotypes/VDJ expression.}
  \item{Principal component analysis (PCA)}{ (\code{Seurat::RunPCA} function)}
  \item{Batch correction using \code{Harmony}}{: covariates given in \code{harmony_vars} will be removed in the Harmony regression step. (Optional, if \code{run.harmony == TRUE})}
  \item{UMAP dimensionality reduction}{: \code{Seurat::RunUMAP}, retaining the top principal components, each of which explain at least \code{var_explained_lim} of the variance.}
  \item{k-neighbor network (kNN) construction}{ (\code{Seurat::FindNeighbors})}
  \item{Define cell clusters based on kNN graph}{ (\code{Seurat::FindClusters})}
}
}
