% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{plotCSRtransitions}
\alias{plotCSRtransitions}
\title{Visualising class-switch recombination (CSR) transitions estimated from the data}
\usage{
plotCSRtransitions(
  csr_transitions = NULL,
  TPTObj = NULL,
  SeuratObj = NULL,
  ighc_count_assay_name = "IGHC",
  return_plot = TRUE,
  significance_threshold = 0.9,
  mask_improbable_csr = TRUE,
  mask_threshold = 1,
  curvature = 0.1,
  arrow = grid::arrow(length = unit(6, "pt"), type = "closed"),
  arrow_colour = NULL,
  bar_colour = NULL
)
}
\arguments{
\item{csr_transitions}{list of parsed transitions from the `prepareCSRtransitions` package. If this is supplied and not NULL, the `TPTObj` and `SeuratObj` arguments will be ignored.}

\item{TPTObj}{List of TPT results, output from the `fitTPT` function. Considered only if `csr_transitions` is NULL.}

\item{SeuratObj}{Seurat object. Considered only if `csr_transitions` is NULL.}

\item{ighc_count_assay_name}{name of assay in SeuratObj which holds the IgH productive/sterile transcript counts. (Default: "IGHC")}

\item{return_plot}{Should the CSR transition plot be returned? If FALSE, a named list of `stationary_distribution` and `flux` will be returned which contains the data frames to be visualised in this plot. (Default: TRUE)}

\item{significance_threshold}{The minimum value in the `TPTObj$significance` matrix for a flux to be shown in the
resulting plot (all fluxes with significance below this value will be removed from the visualisation).}

\item{mask_improbable_csr}{Should isotype combinations which represents improbable Class-switch recombination events (i.e. switching back to an isotype 5' to the current one) be removed from visualisation? (Default: TRUE)}

\item{mask_threshold}{the minimum percentage (max 100) of total flux to be shown in the resulting plot. (Default: 1, i.e.
gross fluxes below 1 will be removed from the visualisation)}

\item{curvature}{amount of curvature of the arrows representing CSR fluxes. (Default: 0.1)}

\item{arrow}{`grid::arrow` object specifying the size and aesthetics of the plotted arrows representing CSR fluxes.}

\item{arrow_colour}{Optional. a column from the 'flux' data frame holding numeric data to be visualised as heat scale in the arrow colours. Useful for
overlaying custom comparisons (e.g. flux differences between Wild-type and Knock-down conditions) onto the plot. If NULL, use the flux amounts to scale
the opaqueness of the black arrows. (Default: NULL)}

\item{bar_colour}{optiona. a column from the 'stationary_distribution' data frame holding categorical information to show stationary distribution from differnet conditions as different coloured bars in the bar plot.
If NULL, only 1 bar will be shown per isotype, in grey. (Default: NULL)}
}
\value{
A ggplot2 object containing a bar plot showing stationary distribution of the isotypes, and arrows
indicating the amount of class-switch recombination events estimated from the fitted transition model
}
\description{
`plotCSRtransitions` visualises the isotype-switching dynamics estimated on the given data using transition models implemented in this package.
}
\details{
`plotCSRtransitions` either takes the output of `prepareCSRtransitions` (argument `csr_transitions`) and generate a
bar plot showing stationary distribution of the isotypes and arrows indicating the amount of fluxes estimated from
the fitted transition model. Alternatively, user can supply the fitted Transition Path Theory object (output of the
`fitTPT` function) as argument `TPTObj`, and the associated `SeuratObj`; in this case the `prepareCSRtransitions` function
will be called internally to parse the transition data for visualisation.
}
