% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{plotCSRtransitions_}
\alias{plotCSRtransitions_}
\title{Workhorse function for visualising class-switch recombination (CSR) transitions}
\usage{
plotCSRtransitions_(
  prepared,
  return_plot = TRUE,
  curvature = 0.1,
  arrow = grid::arrow(length = unit(6, "pt"), type = "closed"),
  arrow_colour = arrow_colour,
  bar_colour = bar_colour
)
}
\arguments{
\item{prepared}{list of parsed transitions from the `prepareCSRtransitions` package. If this is supplied and not NULL, the `TPTObj` and `SeuratObj` arguments will be ignored.}

\item{return_plot}{Should the CSR transition plot be returned? If FALSE, a named list of `stationary_distribution` and `flux` will be returned which contains the data frames to be visualised in this plot. (Default: TRUE)}

\item{curvature}{amount of curvature of the arrows representing CSR fluxes. (Default: 0.1)}

\item{arrow}{`grid::arrow` object specifying the size and aesthetics of the plotted arrows representing CSR fluxes.}

\item{arrow_colour}{Optional. a column from the 'flux' data frame holding numeric data to be visualised as heat scale in the arrow colours. Useful for
overlaying custom comparisons (e.g. flux differences between Wild-type and Knock-down conditions) onto the plot. If NULL, use the flux amounts to scale
the opaqueness of the black arrows. (Default: NULL)}

\item{bar_colour}{optiona. a column from the 'stationary_distribution' data frame holding categorical information to show stationary distribution from differnet conditions as different coloured bars in the bar plot.
If NULL, only 1 bar will be shown per isotype, in grey. (Default: NULL)}
}
\value{
A ggplot2 object containing a bar plot showing stationary distribution of the isotypes, and arrows
indicating the amount of class-switch recombination events estimated from the fitted transition model. If
`return.plot == FALSE`, a list of 2 items containing the 'stationary_distribution' and 'flux' data frames
from `prepared`.
}
\description{
`plotCSRtransitions_` is the workhorse for plotting the CSR transition data. Called
by `plotCSRtransitions` to generate the ggplot2 object.
}
\details{
Called by `plotCSRtransitions` to generate the ggplot2 object.
}
