% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{plotFlux}
\alias{plotFlux}
\title{Visualise fluxes between clusters with a network plot}
\usage{
plotFlux(
  TPTObj,
  SeuratObj,
  significance_threshold = 0.9,
  mask_lower_tri = FALSE,
  mask_threshold = 1,
  new_plot = TRUE
)
}
\arguments{
\item{TPTObj}{List of TPT results, output from the `fitTPT` function.}

\item{SeuratObj}{Seurat object}

\item{significance_threshold}{The minimum value in the `TPTObj$significance` matrix for a flux to be shown in the
resulting plot (all fluxes with significance below this value will be removed from the visualisation).}

\item{mask_lower_tri}{Should fluxes which are considered 'reversed' (i.e. in the lower triangle of the flux matrix)
be masked? (Default: FALSE, i.e. all fluxes in the gross_flux matrix is shown)}

\item{mask_threshold}{the minimum percentage (max 100) of total flux to be shown in the resulting plot. (Default: 1, i.e.
gross fluxes below 1 will be removed from the visualisation)}

\item{new_plot}{Should a stand-alone plot be returned. If FALSE, ggplot2 geoms are returned for adding on top of e.g. Seurat::DimPlot. (Default: FALSE)}
}
\value{
If `new_plot` is TRUE, a network plot where the nodes are placed respecting the positioning of cell clusters in the dimensionality-reduction
projection of the data, and the thickness of edges connecting the nodes signify the flux between pairs of clusters. If `new_plot` is FALSE,
the `ggplot2` geom objects are returned to enable plotting the network graph on top of dimensionality projections (see example, where the UMAP
projection is plotted using the function `dim_plot` implemented in this package).
}
\description{
`plotFlux` plots fluxes between cell clusters in the form of a network plot.
}
\details{
`plotFlux` represents fluxes between cell clusters in a graph, where the thickness of the edges signifies the
amount of fluxes between the given pair of clusters. The placement of clusters respect their placements in the
default dimensionality-reduced projection of the gene expression data stored in the Seurat Object (e.g. the UMAP
projection calculated using Seurat). It can either be a stand-alone plot, or be added on top of UMAP projections (see example).
}
