% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{prepareCSRtransitions}
\alias{prepareCSRtransitions}
\title{Prepare class-switch recombination (CSR) transition data for visualisation}
\usage{
prepareCSRtransitions(
  TPTObj,
  SeuratObj,
  ighc_count_assay_name = "IGHC",
  significance_threshold = 0.9,
  mask_improbable_csr = TRUE,
  mask_threshold = 1
)
}
\arguments{
\item{TPTObj}{List of TPT results, output from the `fitTPT` function.}

\item{SeuratObj}{Seurat object}

\item{ighc_count_assay_name}{name of assay in SeuratObj which holds the IgH productive/sterile transcript counts. (Default: "IGHC")}

\item{significance_threshold}{The minimum value in the `TPTObj$significance` matrix for a flux to be shown in the
resulting plot (all fluxes with significance below this value will be removed from the visualisation).}

\item{mask_improbable_csr}{Should isotype combinations which represents improbable Class-switch recombination events (i.e. switching back to an isotype 5' to the current one) be removed from visualisation? (Default: TRUE)}

\item{mask_threshold}{the minimum percentage (max 100) of total flux to be shown in the resulting plot. (Default: 1, i.e.
gross fluxes below 1 will be removed from the visualisation)}
}
\value{
A named list of three entries:
\describe{
  \item{stationary_distribution}{A data.frame of isotypes, their x/y position in the plot, and their stationary distribution estimated from TPT.}
  \item{flux}{A data.frame listing the fluxes which survive the filtering steps applied (see 'Details'). Their x/y positions in the plot (`from`/`to`) and their gross fluxes are listed.}
  \item{c_genes}{a vector of gene names representing the IgH C genes considered in the model.}
}
This list can be directly passed to the `plotCSRtransitions` function for visualisation.
}
\description{
`prepareCSRtransitions` parse data from `fitTPT` to prepare them for visualising the estimated isotype-switching dynamics.
}
\details{
`prepareCSRtransitions` parses data from the `fitTPT` function for visualising them using `plotCSRtransitions`:
* filter flux matrix by significance: the gross_flux matrix from Transition Path Theory (TPT) is filtered using the significance calculated empirically upon comparison with fluxes generated in randomised models. Only fluxes greater than the randomised fluxes (threshold given by `significance_threshold`, default 0.9) are shown.
* filter flux matrix by magnitude: gross_flux matrix is filtered further such that only those fluxes representing more than a given percentage of total fluxes are shown. Default is 1%.
* remove improbable CSR combinations: isotype switches which are improbable (i.e. switching back to an isotype preceding the current isotype) are removed. This can be turned off by indicating `mask_improbable_csr = FALSE`.
A parsed output is returned to be directly fed into `plotCSRtransitions` to visualise the dynamics of CSR in the given dataset estimated using productive/sterile transcript information.
}
