% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{read_loom_matrices}
\alias{read_loom_matrices}
\title{read loom matrices}
\usage{
read_loom_matrices(file, engine = "hdf5r")
}
\description{
reproduce read.loom.matrices from velocyto.R to avoid problems in installing dependencies
}
