% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{mergeVelocytoWithGEX}
\alias{mergeVelocytoWithGEX}
\title{Combine velocyto loom data with AnnData}
\usage{
mergeVelocytoWithGEX(
  anndata_file,
  loom_file,
  anndata_out_filename,
  conda_env = "scicsr"
)
}
\arguments{
\item{anndata_file}{filename pointing to the AnnData file containing gene expression data.}

\item{loom_file, }{character, name of the new loom file to be written out containing the merged data}

\item{anndata_out_filename}{output filename of the merged AnnData object to be written containing both the velocyto data and the gene expression data.}

\item{conda_env}{character, if not NULL this named conda environment is used to perform the merge.
(Default: 'scicsr'). If NULL, no conda environment will be used, the program assumes the python packages `scanpy` and `scvelo` are installed in the local python)}
}
\value{
a output message indicating success of writing out the merged AnnData object into the file given by `anndata_out_filename`.
}
\description{
`mergeVelocytoWithGEX` merges the velocyto spliced/unspliced gene counts with the AnnData object holding single-cell gene expression data.
This is the preprocessing function before calculating RNA velocity using the python scVelo package and workflow.
}
\details{
`mergeVelocytoWithGEX` uses the R reticulate package to run python commands which merge the velocyto counts with the anndata object.
It is assumed that the cell barcodes in the velocyto loom and the anndata object matches up. Please consult `guessBarcodes` and `combineLoomFiles` first to ensure these barcodes do match up.
}
