% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_repertoire.R
\name{repairBarcode}
\alias{repairBarcode}
\title{Repairing cell barcodes in VDJ data frames to match the Seurat object}
\usage{
repairBarcode(
  vdj_list,
  SeuratObj,
  sample_names,
  seurat_sample_column = "sample_id"
)
}
\arguments{
\item{vdj_list}{a list of data.frame, each holding VDJ data of 1 library}

\item{SeuratObj}{Seurat object.}

\item{sample_names}{a vector of sample names in the data. 'Sample' here refers to sequencing library.}

\item{seurat_sample_column}{Column name in `SeuratObj` metadata where the sample name/ID information is stored.}
}
\value{
A list of vdj data.frames with the cell barcodes in the column 'barcode' repaired to match barcodes in the Seurat object.
}
\description{
`repairBarcode` map samples to their specific barcode suffixes in the Seurat object and alters the cell barcode given in the VDJ data frame to match the Seurat object. This facilitates merging the repertoire data into the Seurat object as additional metadata columns.
}
\details{
The function first establishes a mapping table between samples (given in `seurat_sample_column` in the Seurat object) and their sample-specific cell barcode suffix, and then repairs the barcodes given in the VDJ data frame to match the Seurat object.
}
