% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGHcounts.R
\name{summariseIGHreads}
\alias{summariseIGHreads}
\title{cast data frame IGH counts into a matrix}
\usage{
summariseIGHreads(tb, IGHC_granges)
}
\arguments{
\item{tb}{A data.frame, output from `getIGHreadType`.}

\item{IGHC_granges}{`GenomicRanges` object detailing the positions of IgH C genes in the genome.}
}
\value{
An array of cell barcode by IGH gene type (i.e. S/P/C per C gene)
}
\description{
`summariseIGHreads` converts the data.frame of sterile/productive IGH molecules into a count matrix.
}
\details{
The function does a count of molecules classified to be Sterile (S), Productive (P) or C-only (C) for each isotype,
over molecules per cell barcode and cast this count into a matrix of cell barcodes by IGH molecule type.
}
